requirements.txt
django==4.2.11
djangorestframework==3.15.0
celery==5.3.6
redis==4.4.0
lxml==4.9.2
beautifulsoup4==4.11.1
psycopg2-binary==2.9.9
numpy==1.26.4
